De-coupled Ban List DB operations
