+++ src/main/java/com/rtomyj/skc/dao/implementation/BanListHibernateDao.java
@@ -1,8 +1,8 @@
 package com.rtomyj.skc.dao.implementation;
 
 import com.rtomyj.skc.constant.ErrConstants;
+import com.rtomyj.skc.dao.BanListDao;
 import com.rtomyj.skc.dao.Dao;
-import com.rtomyj.skc.enums.ProductType;
 import com.rtomyj.skc.exception.ErrorType;
 import com.rtomyj.skc.exception.YgoException;
 import com.rtomyj.skc.model.banlist.BanListDate;
@@ -10,17 +10,12 @@ import com.rtomyj.skc.model.banlist.BanListDates;
 import com.rtomyj.skc.model.banlist.CardBanListStatus;
 import com.rtomyj.skc.model.banlist.CardsPreviousBanListStatus;
 import com.rtomyj.skc.model.card.Card;
-import com.rtomyj.skc.model.card.CardBrowseResults;
-import com.rtomyj.skc.model.card.MonsterAssociation;
 import com.rtomyj.skc.model.hibernate.BanListTable;
-import com.rtomyj.skc.model.product.Product;
-import com.rtomyj.skc.model.product.ProductContent;
-import com.rtomyj.skc.model.product.Products;
-import com.rtomyj.skc.model.stats.DatabaseStats;
-import com.rtomyj.skc.model.stats.MonsterTypeStats;
 import lombok.extern.slf4j.Slf4j;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.StopWatch;
@@ -31,16 +26,14 @@ import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 
 
 /**
  * Hibernate implementation of DB DAO interface.
  */
-@Repository("hibernate")
+@Repository("ban-list-hibernate")
 @Slf4j
-public class HibernateDao implements Dao
+public class BanListHibernateDao implements BanListDao
 {
 	private static final String UNSUPPORTED_OPERATION_MESSAGE = "HibernateDao not able to execute method.";
 	@Autowired
@@ -87,15 +80,8 @@ public class HibernateDao implements Dao
 		return null;
 	}
 
-
 	@Override
-	public Card getCardInfo(String cardID) throws YgoException
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
-	@Override
-	public List<Card> getBanListByBanStatus(String date, Status status)
+	public List<Card> getBanListByBanStatus(String date, Dao.Status status)
 	{
 		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
 	}
@@ -115,7 +101,7 @@ public class HibernateDao implements Dao
 		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
 	}
 
-	public List<CardsPreviousBanListStatus> getNewContentOfBanList(String banListDate, Status status)
+	public List<CardsPreviousBanListStatus> getNewContentOfBanList(String banListDate, Dao.Status status)
 	{
 		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
 	}
@@ -125,102 +111,19 @@ public class HibernateDao implements Dao
 		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
 	}
 
-	@Override
-	public String getCardBanListStatusByDate(String cardId, String banListDate)
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
-	@Override
-	public List<Card> searchForCardWithCriteria(final String cardId, final String cardName, final String cardAttribute
-			, final String cardColor, final String monsterType, final int limit, final boolean getBanInfo)
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-	public boolean isValidBanList(final String banListDate)
-	{
-		return false;
-	}
-
-	public List<Product> getAllPackDetails()
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
-	public Products getAllProductsByType(final ProductType productType, final String locale)
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
-	public Map<String, Integer> getProductRarityCount(final String packId)
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
-	public Set<ProductContent> getProductContents(final String packId, final String locale)
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
-	public MonsterTypeStats getMonsterTypeStats(final String cardColor)
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
-	public DatabaseStats getDatabaseStats()
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
-	public Set<Product> getProductDetailsForCard(final String cardId)
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
 	public List<CardBanListStatus> getBanListDetailsForCard(final String cardId)
 	{
 		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
 	}
 
-	public CardBrowseResults getBrowseResults(final Set<String> cardColors, final Set<String> attributeSet, final Set<String> monsterTypeSet, final Set<String> monsterSubTypeSet
-			, final Set<String> monsterLevels, Set<String> monsterRankSet, Set<String> monsterLinkRatingsSet)
-	{
+	@Override
+	public boolean isValidBanList(@NotNull String banListDate) {
 		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
 	}
 
-	public Set<String> getCardColors()
-	{
+	@Nullable
+	@Override
+	public String getCardBanListStatusByDate(@NotNull String cardId, @NotNull String banListDate) {
 		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
 	}
-
-	public Set<String> getMonsterAttributes()
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
-	public Set<String> getMonsterTypes()
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
-	public Set<String> getMonsterSubTypes()
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
-	public Set<MonsterAssociation> getMonsterAssociationField(final String monsterAssociationField)
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
-	public Product getProductInfo(final String productId, final String locale)
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
-	public List<Product> getProductsByLocale(final String locale)
-	{
-		throw new UnsupportedOperationException(UNSUPPORTED_OPERATION_MESSAGE);
-	}
-
 }
