+++ src/main/java/com/rtomyj/skc/service/banlist/BanListDatesService.java
@@ -1,6 +1,6 @@
 package com.rtomyj.skc.service.banlist;
 
-import com.rtomyj.skc.dao.Dao;
+import com.rtomyj.skc.dao.BanListDao;
 import com.rtomyj.skc.exception.YgoException;
 import com.rtomyj.skc.model.banlist.BanListDates;
 import lombok.extern.slf4j.Slf4j;
@@ -18,27 +18,27 @@ public class BanListDatesService
 	/**
 	 * Object used to interface with DB.
 	 */
-	private final Dao dao;
+	private final BanListDao banListDao;
 
 
 	/**
 	 * Create object instance.
-	 * @param dao object used to interface with DB.
+	 * @param banListDao object used to interface with DB.
 	 */
 	@Autowired
-	public BanListDatesService(@Qualifier("hibernate") final Dao dao)
+	public BanListDatesService(@Qualifier("ban-list-hibernate") final BanListDao banListDao)
 	{
-		this.dao = dao;
+		this.banListDao = banListDao;
 	}
 
 
 	/**
-	 * Uses dao helper object to retrieve start dates of all ban lists in the database.
+	 * Uses banListDao helper object to retrieve start dates of all ban lists in the database.
 	 * @return List of BanList objects
 	 */
 	public BanListDates retrieveBanListStartDates() throws YgoException
 	{
-		final BanListDates banListDates = dao.getBanListDates();
+		final BanListDates banListDates = banListDao.getBanListDates();
 		banListDates.setLinks();
 
 		return banListDates;
